- Including mapper to transform from external contact format to internal; - Treating possible null pointers when accessing the headers;

+++ src/main/java/com/erickfelix/contactaggregatorservice/external/KenectLabsContact.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.LocalDateTime;
 
 public record KenectLabsContact (
-        int id,
+        Long id,
         String name,
         String email,
         @JsonProperty("created_at") LocalDateTime createdAt,
